155 refactor: move characteristics schema under schema folder
