+++ schemas/characteristics.ts
@@ -1,22 +1,23 @@
 import { EditIcon } from "@sanity/icons";
+import { defineField, defineType } from "sanity";
 
-export default {
+const characteristics = defineType({
   name: "characteristics",
   title: "Características",
   type: "document",
   icon: EditIcon,
   fields: [
-    {
+    defineField({
       name: "characteristic_name",
       title: "Nombre de la característica",
       type: "string",
-    },
-    {
+    }),
+    defineField({
       name: "characteristics_category",
       title: "A qué categoría pertenece?",
       type: "reference",
       to: [{ type: "characteristics_list" }],
-    },
+    }),
   ],
   preview: {
     select: {
@@ -24,4 +25,6 @@ export default {
       subtitle: "characteristics_category.characteristics_list_name",
     },
   },
-};
+});
+
+export default characteristics;
